Test file moved to test/Gradient and added clad-generated derivatives checks
